Added Chinese path support.
When initializing FolderSync object, setting the third parameter to true can support Chinese. Now, it has been confirmed that the correct FolderObj can be constructed, but due to the coding problem of IDE and system terminal, no further operation can be carried out at the moment, and it is still under test.